DeviceCache to keep data on GPU between operations
Also support histogram of uint64_t (as long long int)
and add GPU implementation of sum operation

+++ src/cubHistogram.cu
@@ -2,7 +2,7 @@
 
 using namespace cub;
 
-template <typename T> void cubHistogram(const T *d_samples, int *d_histogram, int num_levels, T lower_bound, T upper_bound, int64_t N) {
+template <typename T> void cubHistogram(const T *d_samples, unsigned long long int *d_histogram, int num_levels, T lower_bound, T upper_bound, int64_t N) {
   // Determine temporary device storage requirements
   void *d_temp_storage = NULL;
   size_t temp_storage_bytes = 0;
@@ -22,19 +22,19 @@ template <typename T> void cubHistogram(const T *d_samples, int *d_histogram, in
 }
 
 extern "C" {
-void cubHistogram_int32(const int32_t *d_samples, int *d_histogram, int num_levels, int32_t lower_bound, int32_t upper_bound, int64_t N) {
+void cubHistogram_int32(const int32_t *d_samples, unsigned long long int *d_histogram, int num_levels, int32_t lower_bound, int32_t upper_bound, int64_t N) {
   cubHistogram(d_samples, d_histogram, num_levels, lower_bound, upper_bound, N);
 }
 
-void cubHistogram_int64(const int64_t *d_samples, int *d_histogram, int num_levels, int64_t lower_bound, int64_t upper_bound, int64_t N) {
+void cubHistogram_int64(const int64_t *d_samples, unsigned long long int *d_histogram, int num_levels, int64_t lower_bound, int64_t upper_bound, int64_t N) {
   cubHistogram(d_samples, d_histogram, num_levels, lower_bound, upper_bound, N);
 }
 
-void cubHistogram_float(const float *d_samples, int *d_histogram, int num_levels, float lower_bound, float upper_bound, int64_t N) {
+void cubHistogram_float(const float *d_samples, unsigned long long int *d_histogram, int num_levels, float lower_bound, float upper_bound, int64_t N) {
   cubHistogram(d_samples, d_histogram, num_levels, lower_bound, upper_bound, N);
 }
 
-void cubHistogram_double(const double *d_samples, int *d_histogram, int num_levels, double lower_bound, double upper_bound, int64_t N) {
+void cubHistogram_double(const double *d_samples, unsigned long long int *d_histogram, int num_levels, double lower_bound, double upper_bound, int64_t N) {
   cubHistogram(d_samples, d_histogram, num_levels, lower_bound, upper_bound, N);
 }
 }
